requirements.txt
Flask==2.2.5
requests==2.31.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
